Added printSimulatorOutput function
Function to print final results of reservation station, RF, RAT and instruction queue added to the bottom of the file. May need to investigate a good way (other than using -1 as sentinel value) for printing / not printing the RF file information. Also not sure if you can print a boolean value as 0 or 1 by using %u string formatting, will look into this later