Working version (to the best of my knowledge)
Seems to be working after I did a lot of debugging and fixed a bunch of stuff. Added a bunch of DEBUG_MODE stuff for checking information after each step of the process to help figure out what was going on. Only thing we may want to fix other than any other hidden bugs would be that vj and vk print 0 when there is a tag for the qj or qk.